requirements.txt
# Core dependencies
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<1.25.0  # numpy 1.25+ may have issues with some Python versions
tqdm>=4.62.0

# LangChain for agent
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10
langchain-openai>=0.0.5  # If using OpenAI

# NLP
spacy>=3.4.0,<3.8.0  # Tested versions for Python 3.10/3.11

# For local LLM (Ollama)
# Note: Ollama itself must be installed separately
# See: https://ollama.ai/download

# Optional but recommended
python-dotenv>=0.19.0  # For environment variables

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0
ruff>=0.1.0